bug fix to discounts methods

+++ Minimart.Core/Domain/Logic/DiscountOnOrder.cs
@@ -33,7 +33,7 @@ namespace Minimart.Core.Domain.Logic
                 if (this.products.Contains(item.ProductId) && item.Quantity >= _voucher.UnitOrder.Value)
                 {
                     var amount = (item.Price * item.Quantity);
-                    var discount = ((_voucher.Percent ?? 0) / 100); //if Percent = 20, then discount is = 0.2
+                    var discount = ((_voucher.Percent ?? 0) / (decimal)100); //if Percent = 20, then discount is = 0.2
                     item.TotalWithDiscount = amount - (item.Price * discount); //discount on the Nth. unit (1 unit)
                 }
             }
